dma: Use correct len type
vfio_iommu_type1_dirty_bitmap_get.size is of type __u64
dma_controller_dirty_page_get() recieves it as int.

Fixes: #477

+++ lib/dma.c
@@ -482,7 +482,8 @@ dma_controller_dirty_page_logging_stop(dma_controller_t *dma)
 
 int
 dma_controller_dirty_page_get(dma_controller_t *dma, vfu_dma_addr_t addr,
-                              int len, size_t pgsize, size_t size, char **data)
+                              uint64_t len, size_t pgsize, size_t size,
+                              char **data)
 {
     int ret;
     ssize_t bitmap_size;
